메인페이지 평점/별점 수정

+++ src/main/java/kr/or/eum/product/model/vo/ProductAndEtc.java
@@ -18,5 +18,5 @@ public class ProductAndEtc {
 	private int sale;
 	private String productImgPath;
 	private int wishCount;
-	private int revscore;
+	private double revscore;
 }
